passing screenName in props of navbar
